feat: Add option to generate edges for layers and apps
This commit adds the option to generate edges for layers and apps. The `Layer` model now has a new boolean attribute called `generate_edges`, which is used to determine whether or not to generate edges for the layer. Similarly, the `App` model has a new boolean attribute called `generate_layers_edges`, which is used to determine whether or not to generate edges for all layers in the app.

The migration files `2023_10_04_063409_add_generate_edges_to_layers_table.php` and `2023_10_04_063604_add_generate_edges_to_apps_table.php` have been added to create these new columns in the respective tables.

Additionally, a new method called `generateLayerEdges()` has been added to the `Layer` model. This method generates next and previous stages of each track inside the layer and returns them as JSON.

These changes allow users to easily enable or disable edge generation for layers and apps, providing more flexibility in data visualization.