feat: Add show_searchbar toggle to App resource
This commit adds a new toggle field called "show_searchbar" to the App resource in the Nova application. The toggle allows users to show or hide the search bar in the UI. The default value is set to true, and it is hidden from the index view.

Additionally, a new migration file has been created to add the "show_search" column to the "apps" table in the database.