feat: Add related POIs to EcTrack detail view (oc:3034)
This commit adds a new field called "Related POIs" to the EcTrack detail view in the Nova admin panel. The field allows users to select and display associated points of interest (POIs) for a track. It includes functionality for selecting multiple POIs, displaying their names in different languages, and removing selected POIs. The changes include modifications to the `EcTrack.php` file and the addition of a new component called `DetailField.vue`.