feat: Add ExportFormDataXLSX action oc:2807
This commit adds a new file `ExportFormDataXLSX.php` which contains the implementation of the `ExportFormDataXLSX` action. This action allows exporting form data to an XLSX file. The exported data is grouped by app and each app's data is saved in a separate sheet within the XLSX file. The exported data includes field labels and values for each form schema. The generated XLSX file can be downloaded using a temporary signed URL.

The commit also introduces two helper classes, `Sheet` and `MultiSheetExport`, which are used to define the structure and styling of each sheet in the XLSX file.

This feature enhances the functionality of the application by providing users with an easy way to export form data in a structured format for further analysis or reporting purposes.