feat: Add splash screen show option to App resource oc:2556
This commit adds a new toggle field "Show splash screen" to the App resource in Nova. The toggle allows users to enable or disable the display of a splash screen. The default value is set to false and the field is hidden from the index view.

The `ConfTrait` has also been updated to include the `splash_screen_show` property in the data array when saving an App instance.

Additionally, a new migration file has been created (`2024_05_16_145554_add_splash_screen_show_to_apps_table.php`) which adds a `splash_screen_show` column with a boolean type and default value of false to the `apps` table.