chore: Update signature of geohub:update_track_from_osm command oc:2257
The signature of the `geohub:update_track_from_osm` command has been updated to include an optional argument for email addresses. This allows multiple email recipients to be specified when there are errors during the track update process. The emails will be sent using the `UpdateTrackFromOsmEmail` class. Additionally, the format of the error message has been modified to include a link to the OpenStreetMap relation associated with each track.