feat: Add associatedApps relationship to Layer model oc:3474
This commit adds a new method `associatedApps()` to the Layer model, which defines a morphedByMany relationship with the App model. This allows a layer to be associated with multiple apps.

It also includes changes in the Nova resource for Layer, adding an AttachMany field for managing the associated apps. The resource now displays the list of associated apps and allows users to add or remove associations.

A new migration file `2024_06_05_121658_create_app_layer_table.php` is created to create the pivot table `app_layer` for storing the many-to-many relationship between layers and apps.