chore: Update UpdatePOIFromOsm command oc:3476
This commit updates the `UpdatePOIFromOsm` command by adding a new optional flag `--ec_poi_id` to specify the ID of a specific POI to update. If the flag is provided, the command will update only that POI instead of updating all POIs. Additionally, this commit improves the logic for updating feature images by checking if the existing image is up to date before fetching and updating it.