chore: store conf in aws oc:3761
This commit updates the subproject by making changes to the `AppController.php` file, `App` model, and `filesystems.php` configuration file. The changes include refactoring code in the `icon`, `splash`, `iconSmall`, `featureImage`, `iconNotify`, `logoHomepage`, `getOrDownloadIcon`, and `vectorStyle` methods of the controller, adding a new method called `layer`, modifying the logic in the `config` method of the model, and adding a new disk configuration for storing configuration files.