feat(app): ✨ add track direction arrow feature oc:5743 (#85)
* feat(app): ✨ add track direction arrow feature

Added a new toggle option "Show Track Direction Arrow" to the App resource, allowing administrators to enable or disable the track direction arrow on the map. This feature is controlled by a new database column `show_track_direction_arrow`, which has been added to the `apps` table through a migration. The default value for this column is set to true, and it is included in the AppFactory for testing purposes. This enhancement provides better visualization options for map tracking.

* style(app): 💄 improve code readability by adding spaces around concatenation operators

This commit updates the code style by adding spaces around concatenation operators for better readability. Additionally, several comments were added for future enhancements such as including new parameters in the `wmpackage`. These changes ensure that the code is easier to read and maintain.

* Fix styling

* chore(comments): ✏️ update TODO comments to standardized format

Updated all TODO comments across various files to use a consistent format "TODO: ADD2WMPACKAGE". This change ensures uniformity and clarity in the codebase, making it easier to track pending tasks related to the WMPackage integration.

---------

Co-authored-by: peco11 <peco11@users.noreply.github.com>